refactor: move error to okteto pkg
