+++ pkg/cmd/login/browser.go
@@ -22,7 +22,6 @@ import (
 	"net/url"
 
 	"github.com/okteto/okteto/pkg/cmd/login/html"
-	oktetoErrors "github.com/okteto/okteto/pkg/errors"
 	"github.com/okteto/okteto/pkg/okteto"
 	"github.com/okteto/okteto/pkg/types"
 )
@@ -55,7 +54,7 @@ func (h *Handler) handle() http.Handler {
 				return
 			}
 			// we need to return a legible error for the CLI to display
-			h.errChan <- oktetoErrors.TranslateAuthError(err)
+			h.errChan <- okteto.TranslateAuthError(err)
 			return
 		}
 		if err := html.ExecuteSuccess(w); err != nil {
